survey_generator_project/apis_v1/serializers/choice_serializer.py
from rest_framework.serializers import ModelSerializer

from apis_v1.models import ChoiceModel


class ChoiceSerializer(ModelSerializer):
    class Meta:
        model = ChoiceModel
        fields = ['text', 'sequence_no']
